+++ internal/domain/command/clear_export_test.go
@@ -1,0 +1,7 @@
+package command
+
+import "time"
+
+func SetClearClockForTest(h *ClearCommandHandler, now func() time.Time) {
+	h.now = now
+}